dockercon2020-demo/app/requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
# debugging only
# ptvsd==4.3.2
